related-files-apply: remove default implementation
Having a default implementation means that jumpers for which there is
no proper method will be treated as functions, which can result in
errors. This happens when a jumper is called with a place which it
cannot handle (e.g. a recipe jumper being called with a buffer). The
solution is to specialise all the jumper methods to specify what kind
kind of jumper they expect---jumper/place for which there is no
implementation will now just be ignored.