related-files--completing-read: support multiple place categories
Add and document metadata support for displaying more than one type of
place. The api implemented here is not part of the standard Emacs
completion api (which only supports one type per completion session).
Instead, it is a popular way around this limitation, implemented by
Embark, Consult and Marginalia. It is added in major part for
integration with these packages.